share OutputData across functions
